moved and refactored where and how in-house medications constants and types are placed and in-house meds config
