+++ packages/utils/lib/types/data/index.ts
@@ -12,3 +12,4 @@ export * from './paperwork.types';
 export * from './payment';
 export * from './telemed';
 export * from './vitals';
+export * from './medication-administration';
